requirements.txt
asgiref==3.2.7
Django==3.0.5
django-tastypie==0.14.3
Pillow==7.1.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
